requirements.txt
pydantic-ai>=0.0.5
openai>=1.0.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
